requirements.txt
django==5.2.5
djangorestframework==3.16.1
django-tenants==3.8.0
psycopg2-binary==2.9.10
pillow==11.3.0
django-cors-headers==4.7.0
python-decouple==3.8
shortuuid==1.0.11
djangorestframework-simplejwt==5.5.1
django-extensions==4.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
django-storages==1.14.6
boto3==1.40.35